+++ src/lang_utils/error_codes.ml
@@ -269,6 +269,7 @@ let warning_codes = [
   "M0244", None, "Mutable variable is never reassigned";
   "M0254", None, "Initial actor requires field";
   "M0265", None, "system capability not required by this mixin";
+  "M0266", None, "Float32 literal has more precision than Float32 can represent"
   ]
 
 let try_find_explanation code =